requirements.txt
hydra-core==1.3.2
matplotlib==3.8.4
mushroom_classification==0.0.1
numpy==2.0.0
omegaconf==2.3.0
pandas==2.2.2
Pillow==10.4.0
pytorch_lightning==2.3.2
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
timm==1.0.7
torch==2.3.0
torchvision==0.18.0
typer==0.12.3
typing_extensions==4.12.2
wandb==0.17.4
